improve: clone_sources: the function now aims for a shallow clone if possible
This commit is mainly a workaround fix for the randomly failing make_hss function in the
ci. This failure may be caused by the file system used by the ci
machines but the exact reason is not yet determined. By using shallow cloning,
it was found that these fails are prevented.

The updated clone_sources works by aiming to clone the minimum possible from the
source repo's specified in the YAML. If the commit specified isn't found, the depth
of the fetch doubles each attempt until it is found. The .zip and download source
types have been consolidated together. The function itself has been reformatted
to be more readable.

Signed-off-by: Brian Burke <Brian.Burke@microchip.com>