cleanup: shls: remove patches causing unnecessary divergence with SmartHLS projects
The files removed were more for illustration and didn't add any functional value
for the current build option (as only a single SmartHLS project can fit), and may
cause issues with users using the fft function, as the address offset was designed
with the idea of the fft module being added as a second accelerator. The modifications
may also cause unintended consequences for users using SmartHLS specific scripts or setups
using the cloned fpga-hls-libraries repo.

The remaining patches are renamed with correct numbering, and will be removed once the
necessary fixes/renaming occur in the fpga-hls-libraries repo.

Signed-off-by: Brian Burke <Brian.Burke@microchip.com>