fix: ci: pathing to shls projects now works
The sin-shls-apu build-option wasn't generating the specified
design in the ci, and was skipping the integration of the
SmartHLS project. This change fixes the pathing, and throws an
error if an invalid SmartHLS path is passed.

By using the INITIAL_DIRECTORY parameter, the path to the
fpga-hls-libraries source can be resolved for both the
build-bitstream and create-release scripts.

The elf files are now also copied successfully to the
output directory.

Signed-off-by: Brian Burke <Brian.Burke@microchip.com>